+++ dashboard/src/main/java/net/oneandone/stool/dashboard/config/DashboardConfiguration.java
@@ -80,7 +80,7 @@ public class DashboardConfiguration {
         }
         user = user();
         // TODO system.setStoolBin(bin());
-        return Session.load(lib, Logging.create(logs(), "dashboard", user), user, "dashboard", console(), world(), null, svnuser, svnpassword);
+        return Session.load(lib, Logging.create(logs(), "dashboard", user), user, "dashboard", console(), world(), svnuser, svnpassword);
     }
 
     @Bean